+++ src/apis/user/info.py
@@ -27,7 +27,7 @@ async def get_list(request: Request, response: Response, authorization: str = He
     for k in rl[1].keys():
         response.headers[k] = rl[1][k]
 
-    au = await auth(authorization, request, allow_application_token = True, required_permission = ["admin", "hr", "hrm", "get_pending_user_list"])
+    au = await auth(authorization, request, allow_application_token = True, required_permission = ["admin", "hrm", "hr", "get_pending_user_list"])
     if au["error"]:
         response.status_code = au["code"]
         del au["code"]