+++ src/apis/user/manage.py
@@ -260,7 +260,7 @@ async def get_ban_list(request: Request, response: Response, authorization: str 
     for k in rl[1].keys():
         response.headers[k] = rl[1][k]
 
-    au = await auth(authorization, request, allow_application_token = True, required_permission = ["admin", "hr", "hrm", "ban_user"])
+    au = await auth(authorization, request, allow_application_token = True, required_permission = ["admin", "hrm", "hr", "ban_user"])
     if au["error"]:
         response.status_code = au["code"]
         del au["code"]
@@ -312,7 +312,7 @@ async def get_ban(request: Request, response: Response, authorization: str = Hea
     for k in rl[1].keys():
         response.headers[k] = rl[1][k]
 
-    au = await auth(authorization, request, allow_application_token = True, required_permission = ["admin", "hr", "hrm", "ban_user"])
+    au = await auth(authorization, request, allow_application_token = True, required_permission = ["admin", "hrm", "hr", "ban_user"])
     if au["error"]:
         response.status_code = au["code"]
         del au["code"]
@@ -365,7 +365,7 @@ async def put_ban(request: Request, response: Response, authorization: str = Hea
     for k in rl[1].keys():
         response.headers[k] = rl[1][k]
 
-    au = await auth(authorization, request, required_permission = ["admin", "hr", "hrm", "ban_user"])
+    au = await auth(authorization, request, required_permission = ["admin", "hrm", "hr", "ban_user"])
     if au["error"]:
         response.status_code = au["code"]
         del au["code"]
@@ -454,7 +454,7 @@ async def delete_ban(request: Request, response: Response, authorization: str = 
     for k in rl[1].keys():
         response.headers[k] = rl[1][k]
 
-    au = await auth(authorization, request, required_permission = ["admin", "hr", "hrm", "ban_user"])
+    au = await auth(authorization, request, required_permission = ["admin", "hrm", "hr", "ban_user"])
     if au["error"]:
         response.status_code = au["code"]
         del au["code"]