+++ changelog.md
@@ -21,6 +21,7 @@
 17.Added `?after_vehicleid` to **GET** `/economy/trucks/list`  
 18.Added `?after_txid`/`?after`/`?before` to **GET** `/economy/trucks/{vehicleid}/{operation}/history`  
 19.Added **DELETE** `/user/notification` to delete notifications for current user / all users  
+20.Added switch `--enable-performance-header` to add `X-Response-Time` header in response (\*added in most scenarios)  
 
 **v2.5.10**  
 1.Prevented **POST** `/tracksim/update/route` from being loaded when `tracker` plugin is not activated